refactor: 파일명 util 변경 (#86)
